General mapping of most classes
Thanks to research done by @shadowninja108 and myself, most classes have had their member variables added (including placeholders to be replaced later), completing the layout for most, if not all of them.

+++ include/xlink2/xlink2ResAssetCallTable.h
@@ -6,7 +6,7 @@ namespace xlink2 {
 struct ResAssetCallTable {
     u32 keyNamePos;
     s16 assetId;
-    u16 flag;
+    u8 flag;
     s32 duration;
     s32 parentIndex;
     u16 enumIndex;
